feat(httpapi): Adding stubs for an http api implementation.
This commit starts an implementation of an http api
for creating a job to put on a queue for reserving a
recorder and obtaining an access JWT to use for recording purposes.
The http api is implemented in typescript and is currently setup to
have separate npm commands from the existing implementation.